Merget controllers flyttet medarbejderService og medarbejderServiceImpl til service
